backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv==1.0.0
google-generativeai==0.3.1
openai>=1.0.0
httpx==0.25.2
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
aiofiles==23.2.1
PyJWT==2.8.0


